requirements.txt
fastapi~=0.115.0
starlette~=0.41.0
SQLAlchemy~=2.0.36
pydantic-settings~=2.6.0
pydantic~=2.10.0
redis~=5.2.0
python-dotenv~=1.0.1
torch~=2.6.0
pyannote.audio~=3.3.2
openai-whisper~=20240930
psycopg2-binary~=2.9.9
python-multipart~=0.0.12
uvicorn~=0.32.0